% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_intercept_changes.R
\name{extract_change_point_dates}
\alias{extract_change_point_dates}
\title{Function to extract change point dates from 
\code{\link{predict_intercept_changes}}'s output.}
\usage{
extract_change_point_dates(df)
}
\arguments{
\item{df}{Data frame from \code{\link{predict_intercept_changes}}.}
}
\value{
Tibble.
}
\description{
Function to extract change point dates from 
\code{\link{predict_intercept_changes}}'s output.
}
\author{
Stuart K. Grange
}
