% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_intercept_changes.R
\name{extract_intercept_change_values}
\alias{extract_intercept_change_values}
\title{Function to extract intercept change values from 
\code{\link{predict_intercept_changes}}'s output.}
\usage{
extract_intercept_change_values(df)
}
\arguments{
\item{df}{Data frame from \code{\link{predict_intercept_changes}}.}
}
\value{
Tibble.
}
\description{
Function to extract intercept change values from 
\code{\link{predict_intercept_changes}}'s output.
}
\author{
Stuart K. Grange
}
