% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_intercept_changes.R
\name{find_intercept_changes}
\alias{find_intercept_changes}
\title{Function to detect intercept changes in a time series.}
\usage{
find_intercept_changes(
  df,
  n_pieces = 2,
  chains = 3,
  n_cores = 1,
  verbose = FALSE
)
}
\arguments{
\item{df}{Input data frame containing a time series.}

\item{n_pieces}{Number of different pieces in the time series to detect.}

\item{chains}{Number of chains to run.}

\item{n_cores}{Number of cores to be used for model calculation.}

\item{verbose}{Should the function give messages?}
}
\value{
\code{mcpfit} model object with an extra \code{"mcmc_tidy"} element.
}
\description{
Function to detect intercept changes in a time series.
}
\seealso{
\code{\link{predict_intercept_changes}}
}
\author{
Stuart K. Grange
}
