% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_intercept_changes.R
\name{predict_intercept_changes}
\alias{predict_intercept_changes}
\title{Function predict intercept changes from a model created by 
\code{\link{find_intercept_changes}}}
\usage{
predict_intercept_changes(model, n = 20, summarise = FALSE, tz = "UTC")
}
\arguments{
\item{model}{\code{mcpfit} model object from 
\code{\link{find_intercept_changes}}.}

\item{n}{Number of predictions to make.}

\item{summarise}{Should the samples be summarised (the median)?}

\item{tz}{Time zone to parse dates to.}
}
\value{
Tibble.
}
\description{
Function predict intercept changes from a model created by 
\code{\link{find_intercept_changes}}
}
\author{
Stuart K. Grange
}
